refactor: move db, search, skill, hooks verbs to cli commands modules
